feat: implement DashboardInteractive component and restructure page layout
