+++ src/app/layout.tsx
@@ -38,16 +38,29 @@ export default function RootLayout({
       <head>
         {/* Prevent background flash before next-themes hydrates */}
         <style>{`html { background-color: #0d1117; }`}</style>
+        {/* Preconnect to critical origins */}
+        <link rel="preconnect" href="https://fonts.googleapis.com" />
+        <link
+          rel="preconnect"
+          href="https://fonts.gstatic.com"
+          crossOrigin="anonymous"
+        />
+        <link
+          rel="preconnect"
+          href="https://polyfill-library.fastly.dev"
+        />
         {/* Preload the critical above-the-fold logo asset */}
         <link
           rel="preload"
           href="/sf.webp"
           as="image"
           type="image/webp"
+          fetchPriority="high"
         />
         <Script
           id="polyfill-loader"
-          strategy="beforeInteractive"
+          strategy="afterInteractive"
+          fetchPriority="low"
           dangerouslySetInnerHTML={{
             __html: `
               if (!('IntersectionObserver' in window) ||
